pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ansel-agent-backend"
version = "0.1.0"
description = "Minimal backend slice for anselAgent"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "jinja2>=3.1,<4.0",
  "pillow>=10.4,<12.0",
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0,<7.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.3,<5.0",
  "ty>=0.0.1a15,<0.1",
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["server/tests"]

[tool.setuptools.packages.find]
include = ["server*", "shared*"]

[tool.setuptools.package-data]
server = ["codex_bridge/prompts/**/*.txt", "codex_bridge/prompts/**/*.j2", "codex_bridge/data/**/*.yaml"]

[tool.ty.src]
include = ["server", "shared"]
exclude = ["server/tests"]

[tool.ty.environment]
python = ".venv"
python-version = "3.14"

[tool.ty.rules]
invalid-assignment = "ignore"
invalid-return-type = "ignore"
unresolved-attribute = "ignore"
